Editor navbar linked and working
