+++ svgmovestudio.ui/src/components/shared/SVGEditorSidePanel/SVGEditorSidePanel.js
@@ -1,83 +1,33 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Form, FormGroup, Label, Input, InputGroup, InputGroupAddon, InputGroupText, Col, Button } from 'reactstrap';
 import './SVGEditorSidePanel';
+
+import PositionEditor from '../../shared/SVGElementEditors/PositionEditor/PositionEditor';
+import ScaleEditor from '../../shared/SVGElementEditors/ScaleEditor/ScaleEditor';
 
 class SVGEditorSidePanel extends React.Component {
   static propTypes = {
-    createMoveElementObject: PropTypes.func.isRequired 
-  }
-
-  state = {
-    editorControl: {
-      x_position: 0,
-      y_position: 0
-    }
-  }
-
-  moveElementClickEvent = (e) => {
-    const { createMoveElementObject } = this.props;
-    const { editorControl } = this.state;
-    e.preventDefault();
-    createMoveElementObject(editorControl.x_position, editorControl.y_position);
-  }
-
-  x_positionChange = e => {
-    const tempEditorControl = { ...this.state.editorControl };
-    tempEditorControl.x_position = e.target.value;
-    this.setState({ editorControl: tempEditorControl })
-  }
-
-  y_positionChange = e => {
-    const tempEditorControl = { ...this.state.editorControl };
-    tempEditorControl.y_position = e.target.value;
-    this.setState({ editorControl: tempEditorControl })
+    createMoveElementObject: PropTypes.func.isRequired, 
+    selectedEditor: PropTypes.string.isRequired
   }
 
   render() {
-    const { editorControl } = this.state;
+    const { selectedEditor, createMoveElementObject } = this.props;
+
+    const showSelectedEditor = () => {
+      switch(selectedEditor) {
+        case 'position':
+          return <PositionEditor createMoveElementObject={createMoveElementObject}/>
+        case 'scale':
+          return <ScaleEditor />
+        default:
+          break;  
+      }    
+     }
 
     return (
-      <div className="SVGEditorSidePanel border bg-light rounded col-3 m-0">
-        Position
-        <hr/>
-        <Form>
-          <FormGroup row>
-            <Label for="x_position" sm={5}>Horizontal</Label>
-            <Col sm={7}>
-              <InputGroup>
-                <Input 
-                  type="number" 
-                  name="x_position" 
-                  id="x_position" 
-                  value={editorControl.x_position}
-                  onChange={this.x_positionChange} 
-                />
-                <InputGroupAddon addonType="append">
-                  <InputGroupText>px</InputGroupText>
-                </InputGroupAddon>
-              </InputGroup>
-            </Col>         
-          </FormGroup>
-          <FormGroup row>
-            <Label for="y_position" sm={5}>Vertical</Label>
-            <Col sm={7}>
-              <InputGroup>
-                <Input 
-                  type="number" 
-                  name="y_position" 
-                  id="y_position" 
-                  value={editorControl.y_position}
-                  onChange={this.y_positionChange} 
-                />
-                <InputGroupAddon addonType="append">
-                  <InputGroupText>px</InputGroupText>
-                </InputGroupAddon>
-              </InputGroup>
-            </Col>         
-          </FormGroup>
-          <Button className="btn-dark mb-3 w-100" onClick={this.moveElementClickEvent}>Submit Change</Button>
-        </Form>      
+      <div className="SVGEditorSidePanel border bg-light rounded col-3 m-0 pt-2">
+        {showSelectedEditor()}    
       </div>
     )
   }
